+++ src/retry.rs
@@ -1,7 +1,7 @@
 use std::cmp::min;
 
 pub trait RetryPolicy {
-    fn next_delay(&self, attempts: u32) -> i64;
+    fn next_delay(&self, attempts: i16) -> i64;
 }
 
 pub struct ExponentialBackoff {
@@ -15,15 +15,15 @@ pub struct LinearBackoff {
 }
 
 impl RetryPolicy for ExponentialBackoff {
-    fn next_delay(&self, attempts: u32) -> i64 {
+    fn next_delay(&self, attempts: i16) -> i64 {
         let jitter: i64 = rand::random_range(0..10);
-        let delay = self.base * 2i64.pow(attempts) + jitter;
+        let delay = self.base * 2i64.pow(attempts as u32) + jitter;
         min(delay, self.max_delay)
     }
 }
 
 impl RetryPolicy for LinearBackoff {
-    fn next_delay(&self, attempts: u32) -> i64 {
+    fn next_delay(&self, attempts: i16) -> i64 {
         let jitter: i64 = rand::random_range(0..10);
         let delay = self.step * attempts as i64 + jitter;
         min(delay, self.max_delay)
